Add functionality for transmute_header()

+++ zerglib.h
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 static const int OFF_LEN = 75; 
 static const int OFF_VERT = 78; 
 static const int OFF_SEQ = 81;
@@ -5,13 +7,12 @@ static const int OFF_SID = 82;
 static const int OFF_DID = 84; 
 static const int OFF_PAY = 89; 
 
-
 typedef struct { 
-int version;   
-int length; 
-int dest_id; 
-int source_id; 
-int sequence; 
+short version;   
+short length; 
+short dest_id; 
+short source_id; 
+short sequence; 
 char *payload; 
 } psy_data;
